Fix issue #9 -- now handles DST and time zones correctly
Previously DST transitions would not be noticed until the widget's
city was changed, Dashboard was reloaded (e.g. logout/login), or
the widget was removed and re-added when dealing with the analog
display.

The digital display would reflect the DST transitions, but in some
cases some of the digital times would reflect the change either a
day early or a day late.

Additionally locations that had a time zone different from the
current time zone of the system would in some cases compute the
sunrise/sunset/twilight times for either the day before or the day
after the actual date in the selected location.