Sol.js: fix occasional failure to draw during initial load
If there are a lot of widgets on the Dashboard including multiple
copies of the Sol widget, the initial onshow event can end up being
sent before the onload handler is finished running.

Additionally sometimes the initial onshow event can just be dropped
entirely when a lot of widgets are present.

Set the onshow handler earlier to avoid missing the onshow event
and force an onshow event at the end of the load handler just in case.