compact_make_db: exclude covering alt names from database
In some cases there are locations in the data that specify a shortened
alternate name that is the same as the name of a different actual
location.  Since the Sol widget does not (yet) allow choosing from
the list when multiple values are returned by the database lookup,
exclude these alternate names that hide/cover/mask actual locations
of the same name so they can actually be chosen in the widget.