TimeZoneHelper: do not access Address Book on 10.8+
It's extremely unlikely that the Address Book would be accessed
in the first place (only if the last selected city was somehow
not set or was invalid).

Attempting to access the Address Book on 10.8+ presents a
Dialog showing that Dashboard has requested access to Contacts.

Since this is only ever used on rare occasions to initialize the
first city shown, skip trying this if running on 10.8+ to avoid
the ugly dialog.